B/Walker.pm: invoke %Ops only with non-zero $Line
In perl-Devel-Cover.spec, I use the following trick:

	# Undefined subroutine &Devel::Cover::set_first_init_and_end called
	echo 'sub Devel::Cover::set_first_init_and_end{}1' >%buildroot/hack.pm
	%define __spec_autodep_custom_pre export PERL5OPT='-I%buildroot -mhack'

The result is also the dependency on 'perl(hack.pm)', though. :-(

It seems that PERL5OPT and command line switches work on behalf of the
main file, so there is no easy way not to "walk" the code generated by
PERL5OPT.

$ PERL5OPT=-MData::Dumper perl -MO=Deparse -e1
use Data::Dumper;
'???';
-e syntax OK
$

It seems that the best we can do is to check the line number, since
PERL5OPT and command line switches set $Line to 0.  However, it also
seems hard to find the right place for the check in the lower-level
walking code.  I tried to check for $cv->START->line in walk_cv, but
it does not always work well.

So I take another approach - instead of "not walking" the code
generated by PERL5OPT, I simply do not invoke %Ops handlers unless
the $Line is set properly.