B/PerlReq.pm: skip << use overload "0+" => ... >>
This is to skip new warning which is generated for overload->import
when 0+ is overloaded and it is the first in the import list.

For default Exporter::import method, however, version check must
still be performed.

$ perl -e 'use DBI qw(10 hren)'
DBI version 10 required--this is only version 1.616 at /usr/share/perl5/Exporter/Heavy.pm line 120.
BEGIN failed--compilation aborted at -e line 1.
$