B/PerlReq.pm: unify func/method processing via entersub
Instead of handling 'method_named' and 'gv' separately, we now tap
into entersub and then see if it happens to be a method call or a
function call.  So now we have two nice callback tables, one for
methods and another for function names which we want to process.

Handling function calls via entersub (instead of gv) is also required
for the next change (which is not yet committed).  Purportedly, we
need to treat Try::Tiny::try() routine the same way as eval(), but
catch() should be handled like normal code.  Basically, we need to
install some scope/visibility marks in %B::Walker::BlockData for
anoncode; but it is too late for try() gv, since anoncode routine
is already processed/pushed to the stack.

This change also fixes new failing tests.  Apparently ($$OP == $$op)
condition didn't hold with list arguments.  We now have new routine
grok_args() with no such check.  The routine still handles only
leading simple constant arguments (lists are not supported - consider
that lists can be evaluated in scalar context).  So there are still
corner cases.  But overall this is an improvement.

As an example of a "corner case" which is extremely hard to handle
in general case, consider Moose::with() calls.

	with 'My::Role'      => { -version => 0.32 },
	     'My::Otherrole' => { -version => 0.23 };

Arguments passed to with() are actually handled with Data::OptList
module (see Moose/Util.pm).